Refactor: Refactor print tool control (2)
