+++ src/app/ui/components/EditComponent/index.ts
@@ -1,3 +1,5 @@
 import EditComponent from './EditComponent';
 
+export type { EditComponentProps } from './EditComponent';
+
 export default EditComponent;
