+++ src/app/ui/components/EditComponent/styles.styl
